wip on validation, flash messages, error messages
success flash message works great with add post, but register get and edit post have bugs

+++ src/main/java/com/rebekahperkins/website/web/controller/PoemController.java
@@ -3,18 +3,22 @@ package com.rebekahperkins.website.web.controller;
 import com.rebekahperkins.website.domain.Poem;
 import com.rebekahperkins.website.domain.User;
 import com.rebekahperkins.website.service.PoemService;
+import com.rebekahperkins.website.web.FlashMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.security.Principal;
+import javax.validation.Valid;
 
 @Controller
 public class PoemController {
@@ -87,16 +91,35 @@ public class PoemController {
   }
 
   @RequestMapping(value = "/edit", method = RequestMethod.POST, params={"submit"})
-  public String edit (Poem poem, Model model, Principal principal) {
+  public String edit (@Valid Poem poem, Model model, Principal principal, BindingResult bindingResult, RedirectAttributes redirectAttributes) {
+
+    if (bindingResult.hasErrors()) {
+      redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.poem", bindingResult);
+      redirectAttributes.addFlashAttribute("poem", poem);
+      //poem id is null if this is add. otherwise it's edit
+      String addOrEditRedirect = poem.getId() == null ? "/add" : String.format("/cats/%d/edit", poem.getId());
+      return "redirect:" + addOrEditRedirect;
+    }
+
     User loggedInUser = (User)((UsernamePasswordAuthenticationToken)principal).getPrincipal();
-    poem.setSubmittedBy(loggedInUser);
-    poem = poemService.addOrUpdate(poem);
+    if (poem.getId() == null){
+      poem.setSubmittedBy(loggedInUser);
+      poem = poemService.addOrUpdate(poem);
+      redirectAttributes.addFlashAttribute("flash", new FlashMessage("Successfully added", FlashMessage.Status.SUCCESS));
+    } else if (poem.getSubmittedBy().getId() == loggedInUser.getId()){
+      poemService.addOrUpdate(poem);
+      redirectAttributes.addFlashAttribute("flash", new FlashMessage("Successfully edited", FlashMessage.Status.SUCCESS));
+    }
     return "redirect:/cats/" + poem.getId();
   }
 
   @RequestMapping(value = "/edit", method = RequestMethod.POST, params={"delete"})
-  public String delete (Poem poem, Model model) {
-    poemService.delete(poem);
+  public String delete (Poem poem, Model model, Principal principal, RedirectAttributes redirectAttributes) {
+    User loggedInUser = (User)((UsernamePasswordAuthenticationToken)principal).getPrincipal();
+    if (poem.getSubmittedBy().getId() == loggedInUser.getId()) {
+      poemService.delete(poem);
+    }
+    redirectAttributes.addFlashAttribute("flash", new FlashMessage("Successfully deleted", FlashMessage.Status.SUCCESS));
     return "redirect:/cats";
   }
 }
